Planet

Module planetAI

Cargonauts Planet AI Module

This module contains the implementation of the PlanetAI trait for the Cargonauts planet. It defines how the planet handle messages from the Orchestrator and the Explorer.

Each handler is defined as a standalone function to keep the logic modular and clean.

Macros

generate_complex_resource 🔒

Structs

CargonautsPlanet 🔒

Functions

create_planet

Creates a new Planet instance set for be a Cargonauts planet.

handle_combine_resource_request 🔒

This handler processes a request to combine two basic or complex resources into a new complex resource using the planet’s combinator, if energy is available. It returns a PlanetToExplorer::CombineResourceResponse message containing the newly crafted complex resource.

handle_energy_cell_request 🔒

This handler returns an PlanetToExplorer::AvailableEnergyCellResponse message containing the number of currently charged energy cells available on the planet. Since the planet has only one energy cell, the value can only be 0 or 1.

handle_generate_resource_request 🔒

This handler processes a request to generate a basic resource using the planet’s generator, if energy is available. It returns a PlanetToExplorer::GenerateResourceResponse message containing the generated resource.

handle_supported_combination_request 🔒

This handler returns a PlanetToExplorer::SupportedCombinationResponse message that wrap the list of complex resources that the planet can currently generate

handle_supported_resource_request 🔒

This handler returns a PlanetToExplorer::SupportedResourceResponse message that wrap the list of basic resources that the planet can currently generate

log_message 🔒