Planet::planetAI

Function create_planet

pub fn create_planet(
    rx_orchestrator: Receiver<OrchestratorToPlanet>,
    tx_orchestrator: Sender<PlanetToOrchestrator>,
    rx_explorer: Receiver<ExplorerToPlanet>,
    planet_id: u32,
) -> Planet

Creates a new Planet instance set for be a Cargonauts planet.

Parameters

rx_orchestrator: Receiver for messages sent from the orchestrator to the planet (OrchestratorToPlanet).

tx_orchestrator: Sender for messages sent from the planet to the orchestrator (PlanetToOrchestrator).

rx_explorer: Receiver for messages sent from the explorer to the planet (ExplorerToPlanet).

planet_id: Unique identifier assigned to the created planet.

Returns

A fully initialized Planet instance configured with:

Planet type PlanetType::C

A default implementation of CargonautsPlanet AI

Support for generating the basic resource BasicResourceType::Carbon

Support for crafting complex resources: Diamond, Life, AIPartner, Dolphin, Robot, Water

Panics

This function panics if the planet creation fails, such as when:

The provided initialization parameters are invalid

Internal constructor logic returns an error

The panic message is: “Planet creation error!”