Planet::planetAI

Function handle_energy_cell_request

fn handle_energy_cell_request(state: &PlanetState) -> Option<PlanetToExplorer>

This handler returns an PlanetToExplorer::AvailableEnergyCellResponse message containing the number of currently charged energy cells available on the planet. Since the planet has only one energy cell, the value can only be 0 or 1.

Parameters

state: Reference to the planet state

Returns

Some(PlanetToExplorer::AvailableEnergyCellResponse) with available_cells set to:

0: if the energy cell is discharged

1: if the energy cell is charged

Panics

This function does not panic.

Logic

The handler:

Initializes a counter to 0

Accesses the energy cell and increments the counter if it is charged

Wraps the counter inside an PlanetToExplorer::AvailableEnergyCellResponse message and returns it