Planet::planetAI

Function handle_supported_resource_request

fn handle_supported_resource_request(
    generator: &Generator,
) -> Option<PlanetToExplorer>

This handler returns a PlanetToExplorer::SupportedResourceResponse message that wrap the list of basic resources that the planet can currently generate

Parameters

generator: Reference to the planet’s generator

Returns

Option<PlanetToExplorer>

Some(PlanetToExplorer::SupportedResourceResponse) if success.

Panics

This function does not panic.

Logic

The planet can craft basic resources, so the handler:

Get the set of available basic resource from the planet generator

Wrap the set in a PlanetToExplorer::SupportedResourceResponse message and return it